access key in remove report for flow 2

+++ src/main/global/place_event.sh
@@ -2,13 +2,15 @@
 #
 # place_event.sh
 #
-# useage: place_event [event]
+# useage: place_event [event] [optional 'filename']
 #
 # Iterates over all application folders and starts the startup.sh routine.
+# if a filename is given, only folders that contain the filename will be processed.
 
 source $FLOWS_HOME/config.sh
 flow=$1
 event=$2
+fileName=$3
 
 if [ -z "$flow" ] ; then
 	echo "No flow given"
@@ -35,7 +37,9 @@ do
         for fileSet in $na/*
         do
 			if [ -d $fileSet ] ; then
-				echo $(date)>$fileSet/$event
+			    if [[ -z "$filename" ]] || [[ -e $fileSet/$fileName ]] ; then
+			        echo $(date)>$fileSet/$event
+			    fi
 			fi
 		done
 	done
